requirements-dev.txt
pylint>=2.17
pylint-django>=2.6
